fix(customer): give the transactional command handlers real transactions
The last two of the nine real java:S6809 sites from the phase 3.1 triage.

CustomerCommandListener.onCommand is a @KafkaListener method and is not
transactional. It dispatched to handleSegmentResolveRequested and
handleSuppressionAddRequested on `this`, both annotated @Transactional, so
Spring's proxy-based advice was bypassed and neither handler had a transaction.

handleSegmentResolveRequested's own note says why that matters: it "runs in its
own transaction so the reply fact is written to the outbox atomically with the
read that produced it". Without the transaction the resolution and the outbox
reply are separate units, so a requester can be told about a resolution that was
never recorded, or the reply can be lost after the read succeeded.

Moved both to a new CustomerCommandHandlers bean so the dispatch crosses the
proxy, taking the uuidOrNull and enumOrNull payload helpers with them since
nothing else used them.

The third branch, handleOutboxReplayRequested, deliberately stays on the
listener: it declares no transaction and needs none, because it only queues
existing outbox rows for re-publication. Sonar did not flag it, which is useful
negative evidence that the rule is reading annotations rather than pattern-
matching self-calls.
