+++ pos-customer/src/main/java/com/positivity/customer/internal/config/CustomerCommandListener.java
@@ -1,17 +1,10 @@
 package com.positivity.customer.internal.config;
 
-import com.positivity.customer.internal.dto.AddSuppressionRequest;
-import com.positivity.customer.internal.enums.ConsentChangeSource;
-import com.positivity.customer.internal.enums.MarketingChannel;
-import com.positivity.customer.internal.enums.SuppressionReason;
 import com.positivity.customer.service.OutboxReplayService;
-import com.positivity.customer.service.SegmentService;
-import com.positivity.customer.service.SuppressionService;
 import java.time.Clock;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.Locale;
-import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.jspecify.annotations.NonNull;
@@ -21,7 +14,6 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.dao.TransientDataAccessException;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
 import tools.jackson.databind.JsonNode;
 import tools.jackson.databind.ObjectMapper;
 
@@ -72,8 +64,11 @@ public class CustomerCommandListener {
     private final Clock clock;
     private final ObjectMapper objectMapper;
     private final OutboxReplayService outboxReplayService;
-    private final SegmentService segmentService;
-    private final SuppressionService suppressionService;
+    /**
+     * The two transactional handlers, in a separate bean so each gets a real transaction: a
+     * self-call from this class's {@code @KafkaListener} method would bypass the transaction proxy.
+     */
+    private final CustomerCommandHandlers commandHandlers;
 
     @KafkaListener(
             topics = "${pos.customer.kafka.commands-topic:customer.commands.v1}",
@@ -95,11 +90,11 @@ public class CustomerCommandListener {
                 return;
             }
             if (COMMAND_SEGMENT_RESOLVE_REQUESTED.equals(commandType)) {
-                handleSegmentResolveRequested(root);
+                commandHandlers.handleSegmentResolveRequested(root);
                 return;
             }
             if (COMMAND_SUPPRESSION_ADD_REQUESTED.equals(commandType)) {
-                handleSuppressionAddRequested(root);
+                commandHandlers.handleSuppressionAddRequested(root);
                 return;
             }
             log.debug("Ignoring unsupported commandType={} message={}", commandType, message);
@@ -152,89 +147,6 @@ public class CustomerCommandListener {
             return Instant.parse(value);
         } catch (Exception _) {
             log.warn("Malformed payload.{}={} on outbox replay command", field, value);
-            return null;
-        }
-    }
-
-    /**
-     * Resolve a segment and reply with {@code customer.segment.resolved}.
-     *
-     * <p>Runs in its own transaction so the reply fact is written to the outbox atomically with
-     * the read that produced it. A malformed or unknown request is dropped rather than retried:
-     * neither can be fixed by redelivery, and a requester waiting on a reply will time out and
-     * ask again.
-     */
-    @Transactional
-    void handleSegmentResolveRequested(JsonNode root) {
-        JsonNode payload = root.path("payload");
-        UUID requestId = uuidOrNull(payload, "requestId");
-        UUID segmentId = uuidOrNull(payload, "segmentId");
-        if (requestId == null || segmentId == null) {
-            log.warn("Ignoring segment-resolve command missing requestId or segmentId");
-            return;
-        }
-        try {
-            segmentService
-                    .resolveAndPublish(requestId, segmentId)
-                    .ifPresent(count ->
-                            log.info("Resolved segment {} for request {}: {} party(ies)", segmentId, requestId, count));
-        } catch (TransientDataAccessException e) {
-            throw e;
-        } catch (RuntimeException e) {
-            log.warn("Segment {} could not be resolved for request {}: {}", segmentId, requestId, e.getMessage());
-        }
-    }
-
-    /**
-     * Apply a provider-feedback suppression (Story #1150): pos-marketing relays a hard bounce
-     * or spam complaint from the shared platform sender, and the address gets hard-blocked
-     * here — suppression outranks whatever consent the party may still hold.
-     *
-     * <p>The raw address exists only inside this command; {@link SuppressionService#add} stores
-     * a normalized hash plus a masked hint. Malformed commands are dropped, not retried:
-     * redelivery cannot repair a missing address or an unknown enum value.
-     */
-    @Transactional
-    void handleSuppressionAddRequested(JsonNode root) {
-        JsonNode payload = root.path("payload");
-        String address = payload.path("address").stringValue(null);
-        MarketingChannel channel = enumOrNull(MarketingChannel.class, payload, "channel");
-        SuppressionReason reason = enumOrNull(SuppressionReason.class, payload, "reason");
-        if (address == null || address.isBlank() || channel == null || reason == null) {
-            log.warn("Ignoring suppression-add command missing address, channel, or reason");
-            return;
-        }
-        UUID partyId = uuidOrNull(payload, "partyId");
-        var entry = suppressionService.add(
-                new AddSuppressionRequest(channel, address, partyId, reason, ConsentChangeSource.SYSTEM));
-        log.info(
-                "Provider feedback suppressed {} address for party {} (reason {}, suppressionId {})",
-                channel,
-                partyId,
-                reason,
-                entry.suppressionId());
-    }
-
-    private static <E extends Enum<E>> @Nullable E enumOrNull(Class<E> type, JsonNode node, String field) {
-        String value = node.path(field).stringValue(null);
-        if (value == null || value.isBlank()) {
-            return null;
-        }
-        try {
-            return Enum.valueOf(type, value.trim().toUpperCase(Locale.ROOT));
-        } catch (IllegalArgumentException e) {
-            return null;
-        }
-    }
-
-    private static UUID uuidOrNull(JsonNode node, String field) {
-        String value = node.path(field).stringValue(null);
-        if (value == null || value.isBlank()) {
-            return null;
-        }
-        try {
-            return UUID.fromString(value);
-        } catch (IllegalArgumentException e) {
             return null;
         }
     }
